apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.0
requests==2.32.3
google-generativeai==0.8.3
